+++ template-parts/global/site-header.php
@@ -16,7 +16,7 @@
       <div class="flex items-center relative py-4 xl:py-6 xl:space-x-12 transition-all duration-300" id="header-inner">
 
         <div class="site-branding relative z-10 shrink-0">
-          <a href="#" class="block">
+          <a href="/" class="block">
             <div id="logo-container" class="w-24 h-24 md:w-32 md:h-32 xl:w-40 xl:h-40 flex items-center justify-center overflow-hidden transition-all duration-300">
               <img src="<?php echo get_stylesheet_directory_uri() . "/assets/images/logo.png" ?>" alt=" Caravan Industry Victoria" class="w-full h-full object-cover">
             </div>