+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2026-01-20
+### Added
+- Single post template for 'Shows' (`single-shows.php`).
+- Content template for single shows (`template-parts/shows/content-single.php`).
+- Extensive library of Show components in `template-parts/shows/components/`:
+    - `page-header.php`
+    - `visitor-information.php`
+    - `whats-on.php`
+    - `exhibitors.php`
+    - `exhibitor-detail.php`
+    - `exhibitor-special.php`
+    - `map-guide.php`
+    - `gallery.php`
+    - `win.php`
+    - `two-columns-1.php`
+- New styles in `src/css/components.css`.
+
+### Changed
+- Updated `site-navigation.php`.
+
 ## [1.7.0] - 2026-01-20
 ### Added
 - CTA Promo component (`template-parts/components/cta-promo.php`).